app/backend/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
pandas
python-dotenv
